cloud-functions/notification-processor/requirements.txt
requests==2.31.0
google-cloud-logging==3.10.0
